requirements.txt
torch
torchvision
transformers
setuptools
numpy
nltk
Pillow
tqdm
matplotlib
attrs
PyYAML
regex
ftfy
requests
filelock
scikit-image
conf
omegaconf
tensorboard
pycocotools
aiohttp
einops
ftfy
termcolor
toml
tornado
regex
requests
urllib3
packaging
Pillow
PyYAML
wandb
webdataset
corenlp
ipython
scikit-learn